Add StyleRoot component to replace isRoot config
