+++ src/index.js
@@ -2,6 +2,7 @@ import Enhancer from './enhancer';
 import Plugins from './plugins';
 import PrintStyleSheet from './components/print-style-sheet';
 import Style from './components/style';
+import StyleRoot from './components/style-root';
 import getState from './get-state';
 import keyframes from './keyframes';
 import {__clearStateForTests} from './resolve-styles';
@@ -13,6 +14,7 @@ function Radium(ComposedComponent: constructor) {
 Radium.Plugins = Plugins;
 Radium.PrintStyleSheet = PrintStyleSheet;
 Radium.Style = Style;
+Radium.StyleRoot = StyleRoot;
 Radium.getState = getState;
 Radium.keyframes = keyframes;
 Radium.__clearStateForTests = __clearStateForTests;
