+++ src/resolve-styles.js
@@ -213,7 +213,8 @@ const _runPlugins = function({
     if (!styleKeeper) {
       throw new Error(
         'To use plugins requiring `addCSS` (e.g. keyframes, media queries), ' +
-          'please add `isRoot: true` to your root component\'s Radium config.',
+          'please wrap your application in the StyleRoot component. Component ' +
+          'name: `' + componentName + '`.',
       );
     }
